requirements.txt
Flask>=2.3.0,<2.4
flask-cors==3.0.10
pymongo==4.4.0
dnspython>=2.8.0,<3.0
PyJWT==2.8.0
bcrypt==4.1.1
python-dotenv==1.0.0
